Fixed incorrect override that disabled stepper fan.
I've added @Override annotations to all overrides.  This revealed
a bug: the enableMotor/disableMotor, despite being documented as
overrides, weren't.  As a result, the Sanguino3GDriver behavior
was being invoked, generating DC motor control messages even when
the axis has been hijacked for a stepper extruder.  Because this
driver reuses the same motor control channel for the stepper fan,
this meant the fan turned on...until the first motion.  Then it
started toggling on and off with the stepper motor.

That's kind of silly, so with this change it obeys the flag passed
to enableStepperExtruderFan(boolean).