@Deprecated/@Override cleanup in Driver classes.
I've deprecated some new methods on Driver.  These methods -- or,
more specifically, the fact that two versions of each of these
methods existed -- hid the stepper fan bug I fixed in a previous
commit.  Interestingly, many of the Driver implementations
deprecated these methods, but the annotations never got propagated
back to the interface, and so some were missed.

At the same time, I added @Override where appropriate to a subset
of the Driver classes (those I use).  The interesting thing about
that change is the methods that *didn't* get @Override.  Driver
defines no multiple-tool-aware version of e.g. getMotorRPM, but
some of the drivers attempt to override it...resulting in dead
code.  To be fixed in a subsequent commit.