Cleaned up DriverBaseImplementation, mostly.
Made it clear that DriverBaseImplementation is not to be used
on its own by marking it abstract and making the constructor
protected.

Put big scary comments on methods that appear to have been
attempts at overrides, but which actually overload methods in
Driver.  Because the rest of RepG manipulates Drivers through
the Driver interface -- not as objects of type
DriverBaseImplementation -- this means all these functions are
dead!  Which is too bad, because they're awesome.  Marked them
for eventual transfer to Driver.

Removed all commented-out code and fields.

Tightened access modifiers as far as possible (e.g. fields to
private, methods to protected).  This revealed that some code
and fields were dead, but not being detected as such because
they were declared as visible; removed all that.

Routed all internal setting of currentPosition through the
setInternalPosition method, which appears to have been provided
for this purpose.  Inlining it would also be a reasonable choice.

Propagated @Deprecated annotations forward from Driver.

There's still some stuff left to do here, but the change does too
many things already.