Modernizing Driver's API.
ReplicatorG is lurching toward supporting multiple toolheads.
The Sanguino3G driver, in particular, has versions of methods
like readToolStatus or getMotorSpeedPWM that take an explicit
toolhead index.

Unfortunately, as shown by my previous commit that moved those
methods to private, nothing ever called them.  This is because
the rest of the code uses the Driver interface, which has only
*some* of these methods.  The migration was partial.

This commit finishes it, with the exception of some little-used
features that I don't understand well enough to test (collets,
for one).  All Driver methods that apply to toolheads now take
an explicit index.  The old versions are still here, but they're
explicitly @Deprecated, so I can easily find all callers to
modernize them, too.

In a surprising bit of honesty, the code describes the practice
of interpreting tool -1 to indicate "current tool" as a "fast
hack to get software out."  I am perpetuating the fast hack in
this commit: all the deprecated methods forward to the modern
ones, but with a tool index of -1.  I will fix this in a
subsequent change.

Note that the RepRap drivers are not fully prepared to support
multitool, and this commit does not change that.