Modernized DriverQueryInterface.
While the Driver interface grew new methods to support multiple
tools, DriverQueryInterface was left behind.  Fortunately, it's
a much less popular type, so fixing this is a relatively small
change.  Because classes tend to implement both Driver and
DriverQueryInterface, and Driver already had the methods in
question, I was able to simply remove the old-style methods
from DriverQueryInterface.

As a side effect, a few G-Code commands are now multi-tool
aware -- specifically, the commands for setting spindle speed.