Reworked GCodeCommand parameter API.
The previous version implicitly returned a magic number that is also
a reasonable parameter value for most parameters: -1.  A comment in
the code ("what do we return?") suggests that this was done with some
reluctance.

There are cases where the magic negative number can mask problems.
For example, the current code will accept a G97 (set spindle RPM)
code without the S parameter.  It gets a -1 for the spindle speed,
and winds up sending a small negative number in the Sanguino3GDriver.
The Sanguino3G firmware uses an unsigned integer for the same field,
so the negative number becomes a large positive number...and so
on.  It's bad practice to use magic return values like this to
indicate errors in-band.

This changes splits that API into two halves.  getCodeValue(char)
now throws if the code isn't present -- it's expected to be used in
concert with hasCode(char).  For cases where we really do want a
magic number returned for nonexistent codes, there's a new method,
getCodeValue(char, double) that specifies the value to return if the
code isn't present.

(I've also added a version of getCodeValueInt from my previous commit
that takes a fallback parameter.)

The refactoring process was as follows:

First, replace all calls to getCodeValue(x) with getCodeValue(x, -1).
This makes the presence of the magic number explicit at all callsites.
In particular, this makes it clear which callsites aren't expecting it,
so we can fix them.

Second, locate all cases where a call to getCodeValue(x, y) is guarded
by a call to hasCode(x).  In these cases, no fallback value is needed,
because the code is present.  Replace all such calls with straight-up
getCodeValue(x).

Third, locate all cases where a value is being conditionally overwritten
by the results of getCodeValue(x).  For example,

int x = DEFAULT_VALUE;
if (command.hasCode('X')) x = command.getCodeValue('X');

Replace all such instances with getCodeValue(x, default).

Finally, locate all cases where the result of getCodeValue is being
cast to an int.  Replace them with getCodeValueInt.

The result of all this: callsites that were aware of the "-1 means
undefined" convention, and guarded their calls to getCodeValue with
hasCode, are unchanged in the end.  From the size of this change, you
can see how few cases this covers.